platform/backend/requirements.txt
fastapi>=0.110
uvicorn>=0.27
pydantic>=2.0
pydantic-settings>=2.0
supabase>=2.0
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
stripe>=8.0
httpx>=0.27
python-multipart>=0.0.9
loguru>=0.7
alembic>=1.13
asyncpg>=0.29
sqlalchemy>=2.0
python-dotenv>=1.0
